+++ app/decorators/models/solidus_easypost/spree/address_decorator.rb
@@ -2,27 +2,25 @@
 
 module SolidusEasypost
   module Spree
-    module EasyPost
-      module AddressDecorator
-        def easypost_address
-          attributes = {
-            street1: address1,
-            street2: address2,
-            city: city,
-            zip: zipcode,
-            phone: phone
-          }
+    module AddressDecorator
+      def easypost_address
+        attributes = {
+          street1: address1,
+          street2: address2,
+          city: city,
+          zip: zipcode,
+          phone: phone
+        }
 
-          attributes[:company] = company if respond_to?(:company)
-          attributes[:name] = full_name if respond_to?(:full_name)
-          attributes[:state] = state ? state.abbr : state_name
-          attributes[:country] = country&.iso
+        attributes[:company] = company if respond_to?(:company)
+        attributes[:name] = full_name if respond_to?(:full_name)
+        attributes[:state] = state ? state.abbr : state_name
+        attributes[:country] = country&.iso
 
-          ::EasyPost::Address.create attributes
-        end
-
-        ::Spree::Address.prepend self
+        ::EasyPost::Address.create attributes
       end
+
+      ::Spree::Address.prepend self
     end
   end
 end
