+++ app/decorators/models/solidus_easypost/spree/stock_location_decorator.rb
@@ -2,27 +2,25 @@
 
 module SolidusEasypost
   module Spree
-    module EasyPost
-      module StockLocationDecorator
-        def easypost_address
-          attributes = {
-            street1: address1,
-            street2: address2,
-            city: city,
-            zip: zipcode,
-            phone: phone,
-            name: name,
-            company: name
-          }
+    module StockLocationDecorator
+      def easypost_address
+        attributes = {
+          street1: address1,
+          street2: address2,
+          city: city,
+          zip: zipcode,
+          phone: phone,
+          name: name,
+          company: name
+        }
 
-          attributes[:state] = state ? state.abbr : state_name
-          attributes[:country] = country&.iso
+        attributes[:state] = state ? state.abbr : state_name
+        attributes[:country] = country&.iso
 
-          ::EasyPost::Address.create attributes
-        end
-
-        ::Spree::StockLocation.prepend self
+        ::EasyPost::Address.create attributes
       end
+
+      ::Spree::StockLocation.prepend self
     end
   end
 end
