+++ spec/solidus_easypost_spec.rb
@@ -2,12 +2,12 @@
 
 require 'spec_helper'
 
-RSpec.describe Solidus::EasyPost do
+RSpec.describe SolidusEasypost do
   describe 'CONFIGS' do
     it 'can be altered' do
       expect {
-        Solidus::EasyPost::CONFIGS[:purchase_labels?] = false
-      }.to change { Solidus::EasyPost::CONFIGS[:purchase_labels?] }.to(false)
+        SolidusEasypost::CONFIGS[:purchase_labels?] = false
+      }.to change { SolidusEasypost::CONFIGS[:purchase_labels?] }.to(false)
     end
   end
 end
